had an idea of using nexus theory to start looking of gods as mortals with advance technologies
